Fix discard page

+++ page_discard.h
@@ -1,6 +1,7 @@
 #ifndef UNTITLED_PAGE_DISCARD_H
 #define UNTITLED_PAGE_DISCARD_H
-void discard_empty_page(void *page_address, size_t page_size) ;
+
+void discard_empty_page(void *page_address, int page_size);
 
 
 #endif //UNTITLED_PAGE_DISCARD_H
